Re-organized Home screen files/folders
