+++ lib/state_management/changelog/changelog_notifier.dart
@@ -1,5 +1,5 @@
 import 'package:anonaddy/global_providers.dart';
-import 'package:anonaddy/screens/home_screen_components/changelog_widget.dart';
+import 'package:anonaddy/screens/home_screen/components/changelog_widget.dart';
 import 'package:anonaddy/services/changelog_service/changelog_service.dart';
 import 'package:anonaddy/shared_components/constants/material_constants.dart';
 import 'package:flutter/material.dart';
